fix : comments delete error

+++ backend/backend-java/src/main/java/com/ssafy/api/service/GoodsServiceImpl.java
@@ -93,8 +93,8 @@ public class GoodsServiceImpl implements GoodsService {
     }
 
     @Override
-    public void goodsCommentDelete(long commentSeq) {
-        goodsReviewRepository.deleteById(commentSeq);
+    public void goodsCommentDelete(int commentSeq) {
+        goodsReviewRepository.deleteById((long) commentSeq);
 
     }
